followup-reminders

+++ database/migrations/2023_07_04_072714_create_transport_allowances_table.php
@@ -29,8 +29,8 @@ return new class extends Migration
             $table->foreignId('created_by')->index()->nullable()->constrained('users')->onDelete('set null');
             $table->foreignId('client_id')->index()->nullable()->constrained('clients')->onDelete('set null');
             $table->foreignId('follow_up_id')->index()->nullable()->constrained('follow_up_infos')->onDelete('set null');
-            $table->integer('allowance_status')->default(0)->comment('0 = pending, 1 = confirmed, 2 = rejected, 3 = warning');
-            $table->integer('travel_status')->default(0)->comment('0 = pending, 1 = confirmed');
+            $table->tinyInteger('allowance_status')->default(0)->comment('0 = pending, 1 = confirmed, 2 = rejected, 3 = warning');
+            $table->tinyInteger('travel_status')->default(0)->comment('0 = pending, 1 = confirmed');
             $table->tinyInteger('is_paid', false, true)->default(0);
             $table->timestamps();
         });
